review: worst ordering
we can move the creation of the orderedODS object outside of the if statements, making the code much more concise. We are able to do this because in the first else statement, we can set the hospital rank for "worst" equal to the number of rows in the orderedODS object (aka the last row). 

that will give us the worst hospital, without having to do any reordering.